bk: 48addcf06JBjSPalT3D0gBp48ibLNw

+++ generic/Last.h
@@ -140,7 +140,7 @@ struct L_expr {
 	union {
 		char *string;
 		double flote;
-		int integer;
+		long integer;
 	} u;
 };
